rearranged to make other grids easier to integrate with or without fem
